+++ src/test/resources/org/bytedeco/javacpp/EnumTest.h
@@ -1,3 +1,7 @@
+enum class BoolEnum : bool {
+    BOOL_ENUM = 1
+};
+
 enum class CharEnum : char {
     CHAR_ENUM = 42
 };
@@ -13,6 +17,10 @@ enum /* no class */ IntEnum : int {
 enum /* no class */ LongEnum : long long {
     LONG_ENUM = 121110
 };
+
+BoolEnum Char2Bool(CharEnum e) {
+    return (BoolEnum)e;
+}
 
 ShortEnum Char2Short(CharEnum e) {
     return (ShortEnum)e;